move the vat files for the 'contractHost' tests to test/swingsetTests/scaffolding
